api/streamlit/requirements.txt
streamlit
requests
Pillow
pandas
python-dotenv